connect: move EE credentials to lib/connect
The handling of User Credentials for connecting to EE has been moved
to the `lib/connect` package. This had little to do with
the installation process, and secondly it created a dependency include
loop between `cli/install_ee` and `cli/search`.

Part of #TNTP-1095
