+++ lib/connect/go.mod
@@ -5,11 +5,13 @@ go 1.23.8
 require (
 	github.com/pmezard/go-difflib v1.0.0
 	github.com/stretchr/testify v1.10.0
+	github.com/tarantool/tt v1.3.1
+	golang.org/x/term v0.13.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
-	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
+	github.com/kr/pretty v0.2.1 // indirect
+	golang.org/x/sys v0.13.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
